+++ src/resources/extensions/gsd/tests/worktree-sync-overwrite-loop.test.ts
@@ -29,7 +29,7 @@ import {
 import { join } from "node:path";
 import { tmpdir } from "node:os";
 
-import { syncProjectRootToWorktree } from "../auto-worktree-sync.ts";
+import { syncProjectRootToWorktree } from "../auto-worktree.ts";
 import { createTestContext } from "./test-helpers.ts";
 
 const { assertTrue, assertEq, report } = createTestContext();
